内部実装に関するファイルを、API 利用に関するものと区別しやすいようにフォルダーで分離しました。
