+++ Sources/YumemiWeather/InternalValueRepresentations/AreaRequest.swift
