+++ Sources/YumemiWeather/InternalValueRepresentations/AreaResponse.swift
