+++ Sources/YumemiWeather/InternalValueRepresentations/Request.swift
