+++ Sources/YumemiWeather/InternalValueRepresentations/WeatherCondition.swift
@@ -1,0 +1,10 @@
+//
+//  WeatherCondition.swift
+//
+
+enum WeatherCondition: String, CaseIterable {
+    case sunny
+    case cloudy
+    case rainy
+}
+
