+++ Sources/YumemiWeather/RandomizationFeatures/YumemiWeather.APIQuality.swift
